pBot/requirements.txt
flask
urllib3
google-generativeai==0.2.1
vertexai
pangea-sdk